renamed UserName/UserNameImpl to UserNameRequestService/UserNameRequestServiceImpl
