+++ spring-shell.log
@@ -43,3 +43,8 @@
 1547112302624:start
 1547112348164:language ru
 1547112350577:start
+1547112739342:language ru
+1547112777142:start
+1547112884539:locale ru
+1547112893422:language ru
+1547112909965:start
