--- src/main/java/ru/otus/spring/hw01/services/UserName.java
@@ -1,5 +1,0 @@
-package ru.otus.spring.hw01.services;
-
-public interface UserName {
-    String getUserName();
-}
