+++ src/main/java/ru/otus/spring/hw01/services/UserNameRequestService.java
@@ -1,0 +1,5 @@
+package ru.otus.spring.hw01.services;
+
+public interface UserNameRequestService {
+    String getUserName();
+}
